refactor: move SimulationNotificationListener in adapter
