+++ src/main/scala/ulisse/adapters/output/SimulationNotificationAdapter.scala
@@ -3,7 +3,6 @@ package ulisse.adapters.output
 import ulisse.applications.ports.SimulationPorts
 import ulisse.entities.simulation.data.SimulationData
 import ulisse.entities.station.Station
-import ulisse.infrastructures.view.simulation.SimulationNotificationListener
 
 /** Adapter for the SimulationPorts.Output port. */
 final case class SimulationNotificationAdapter(private val simulationPage: SimulationNotificationListener*)
