+++ src/main/scala/ulisse/adapters/output/SimulationNotificationListener.scala
@@ -1,0 +1,12 @@
+package ulisse.adapters.output
+
+import ulisse.entities.simulation.data.SimulationData
+
+/** Listener for simulation notifications. */
+trait SimulationNotificationListener:
+
+  /** Update the simulation data. */
+  def updateData(data: SimulationData): Unit
+
+  /** End the simulation. */
+  def endSimulation(data: SimulationData): Unit
